+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 CHANGELOG
 =========
 
-0.2.0 (2014-08-11)
+Next Release (TBD)
 ------------------
 
 Adding new commands and changing how parsing works.
@@ -10,6 +10,8 @@ Adding new commands and changing how parsing works.
 * Made ``TAG`` a required argument for ``parse`` and ``tag``.
 * ``parse`` now only outputs the contents of the changelog entry to STDOUT.
 * Added ``list`` and ``latest`` commands.
+* Added an ``update`` command that can be used to update a CHANGELOG TBD
+  heading with a new changelog entry.
 
 0.1.0 (2014-08-11)
 ------------------